+++ packages/terra-list/tests/wdio/list-spec.js
@@ -96,4 +96,13 @@ Terra.describeViewports('List', ['medium'], () => {
       Terra.validates.element('After Drop Focus Next Item', { selector: '#root' });
     });
   });
+
+  describe('Slidepanel ListItem Drag and Drop', () => {
+    before(() => browser.url('raw/tests/cerner-terra-core-docs/list/slide-panel-draggable-list'));
+
+    it('should perform drag and drop operation', () => {
+      browser.keys(['Tab', 'Space', 'ArrowDown', 'Space']);
+      Terra.validates.element('Slide Panel After Drag and Drop', { selector: '#root' });
+    });
+  });
 });