+++ packages/terra-list/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Fixed
+  * Fixed position loss issue while dragging list item for `terra-list` inside `terra-slide-panel`.
+
 ## 4.62.0 - (September 21, 2023)
 
 * Changed